+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Replacing pkg_resources with the builtin importlib.metadata [687](https://github.com/scanapi/scanapi/issues/687)
 
 ## [2.10.2] - 2024-11-10
 ### Removed